pyproject.toml
[tool.poetry]
name = "docker-postgres"
version = "0.1.0"
description = ""
authors = ["Nicolas Doege <github.repsu@passinbox.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pre-commit = "^4.1.0"
streamlit = "^1.42.0"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
